Extracted GameProvider
